实现点击“回复”后自动登录并在登录成功后跳转到首页

+++ src/main/java/life/majiang/community/controller/PublishController.java
@@ -25,7 +25,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")    //通过下面的PathVariable获得前端传过来的id
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title", question.getTitle());
@@ -44,7 +44,7 @@ public class PublishController {
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "id", required = false) Integer id,
+            @RequestParam(value = "id", required = false) Long id,
             HttpServletRequest request,
             Model model) {
         model.addAttribute("title", title);
@@ -62,8 +62,6 @@ public class PublishController {
             model.addAttribute("error", "标签不能为空");
             return "publish";
         }
-
-
         User user = (User) request.getSession().getAttribute("user");   //注意是"user" 而不是 "name"
         if (user == null) {
             model.addAttribute("error", "用户未登录！");
